Rename package 'entities' to 'domain'
