+++ tests/integration/test_datajoint_persistence.py
@@ -17,8 +17,8 @@ import pytest
 from dj_link.adapters import PrimaryKey
 from dj_link.adapters.gateway import DJLinkGateway
 from dj_link.adapters.identification import IdentificationTranslator
-from dj_link.entities.link import create_link, delete, process, pull
-from dj_link.entities.state import Components, Processes
+from dj_link.domain.link import create_link, delete, process, pull
+from dj_link.domain.state import Components, Processes
 from dj_link.infrastructure.facade import DJLinkFacade, Table
