+++ tests/unit/entities/test_link.py
@@ -5,9 +5,9 @@ from typing import ContextManager, Iterable, Mapping
 
 import pytest
 
-from dj_link.entities.custom_types import Identifier
-from dj_link.entities.link import Link, create_link, delete, process, pull
-from dj_link.entities.state import Components, Processes, State, states
+from dj_link.domain.custom_types import Identifier
+from dj_link.domain.link import Link, create_link, delete, process, pull
+from dj_link.domain.state import Components, Processes, State, states
 from tests.assignments import create_assignments, create_identifier, create_identifiers
